+++ lib/pidgin2adium.rb
@@ -1,5 +1,10 @@
+require 'pathname'
+require 'optparse'
+
 require 'pipio'
+
 require 'pidgin2adium/version'
+require 'pidgin2adium/cli'
 require 'pidgin2adium/file_finder'
 require 'pidgin2adium/runner'
 require 'pidgin2adium/adium_chat_file_creator'